Refactor for smaller generated syntax.
This allows a macro to avoid creating an entire discovery
document as a huge literal (which includes strings for
documentation and other information not actually
needed). Instead, extract only what's necessary into a
`method-spec' struct, which can be passed to a function
that creates the function wrapper at run-time.

Although this works, and the generated syntax is more
compact, the transformer function to generate such
syntax is much more complicated than I expected. I'm
concerned I'm missing a simpler way to go about this.